contact_notes_api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.15.0
pyjwt==2.1.0
sqlalchemy==1.4.23
python-dotenv>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt==3.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
requests==2.26.0
pytest==6.2.5
httpx==0.19.0
email-validator